RAVE 0.434
- Afegits butons al Thumbnail de la barra de tasques.
- Solucionat bug dels scrolls en la llista i l'arbre
- Solucionat bug del MedioActual -1 quan era unsigned...
- El VLC cambia la recta de Clip pel tumbnail quan s'executa un video. Ara al acabar el video torno a posar el Clip per tota la finestra.
- Solucionat bug en el hook global del teclat que no tenia en compte les tecles Control, Shift, Alt.

+++ Reproductor de Audio y Video Extendido/ListaMedios.cpp
@@ -161,7 +161,7 @@ const BOOL ListaMedios::Mezclar(const BOOL nMezclar) {
 		// Busco la posición del medio actual en la lista mezclada
 		for (i = 0; i < _Items.size(); i++) {
 			if (_Items[i] == _MediosOrdenados[MedioActual]) {
-				MedioActual = i;
+				MedioActual = static_cast<long>(i);
 				MostrarItem(i);
 				break;
 			}
@@ -173,7 +173,7 @@ const BOOL ListaMedios::Mezclar(const BOOL nMezclar) {
 		// Busco la posición del medio actual en la lista original
 		for (i = 0; i < _Items.size(); i++) {
 			if (_Items[MedioActual] == _MediosOrdenados[i]) {
-				MedioActual = i;
+				MedioActual = static_cast<long>(i);
 				MostrarItem(i);
 				break;
 			}
